Fix types like `notes: str = None` to have `Optional[str]`
PEP 484 made the Optional type explicit. The mypy tool (static type
checker) started enforcing this rule in newer versions. This commit
fixes all reported-by-mypy errors of this type, like this:

    `notes: str = None` -> `notes: Optional[str] = None`

Note that mypy checker still fails because of the other error types;
they will be fixed in follow up commits.

This commit also adds the installation of types-requests module, which
is a dependency of mypy.

Signed-off-by: Dmitrii Kuvaiskii <dimakuv@amazon.de>