Fix all pylint errors
Newer versions of pylint report more errors on the codebase. Also, we
enabled all error types on pytests under test/, and fixed all issues in
the tests (this may be considered an overkill, but it's easier to fix
the issues now and keep the codebase clean in the future, including the
tests).

As a side effect, a useless created-by-default test file
test/test_package_validation_tool.py was removed. We also noticed
missing license information in test/test_changelog.py and added it.

Signed-off-by: Dmitrii Kuvaiskii <dimakuv@amazon.de>