utils: Use `User-Agent: Wget/1.14 (linux-gnu)` when downloading files
Previously, our tool pretended to be the Firefox web browser to websites
when downloading files (in particular source tarballs). This was
achieved by setting user-agent HTTP header field to "Mozilla/5.0
Firefox/140.0". This is generally wrong: ideally the tool must state
its own name, not pretend to be something else; however in this case
some websites refuse to allow the download. But more practically,
pretending to be a Firefox web browser results in some websites
redirecting to an interactive web page, instead of just downloading the
requested file. This was detected on the OpenSSH tarball URL:
http://pamsshagentauth/pam_ssh_agent_auth.

To stop such behavior of some web sites, this commit stops the practice
of pretending to be a web browser and instead pretends to be a wget
utility.

Signed-off-by: Dmitrii Kuvaiskii <dimakuv@amazon.de>