[suggesting] Use name of source package if archive has only version
Previously, the "suggesting repos" logic relied on the basename of the
local archive to contain both the software project name and its version,
for example "zlib-1.2.11.tar.gz". Sometimes the archive doesn't contain
the project name but only the version, for example "v2.2.0.tar.gz"
(e.g., this is typically how GitHub publishes release artifacts). In
such cases, the tool searched for a bogus upstream repo "v" and failed
to find any matches.

This commit fixes this corner case by using the name of the source
package instead of the archive when calling suggestion methods, if the
local archive seems to have only the version. For example, if local
archive is "v2.2.0.tar.gz", then the source package name ("ec2-utils")
is used instead.

This commit also modifies the "suggesting archives" logic in the same
way. However, there are no suggestion methods that would need this
currently, so this is purely for future extensibility.

Signed-off-by: Dmitrii Kuvaiskii <dimakuv@amazon.de>