matching: save differing files on disk
Previously, when performing local archive vs upstream archive/repo file
matching, a helper function _collect_file_match_statistics() had a stub
to save differing files in a user-specified directory (but only if the
user explicitly set environment variable PVT_FILE_MATCHER_DIFFS_PATH).

This commit replaces the stub (the comment) with the actual
implementation. Note that previously it was the stub because some time
ago  the "local archive vs upstream archive" file matching code operated
on tarballs instead of individual files, so if the tarballs were
different, the file-saving code would only save the tarballs instead of
individual files. This wasn't useful and that's what the comment meant.
This "tarball matching" was already fixed to always extract the
tarballs, so now we can add the file-saving logic easily.

Signed-off-by: Dmitrii Kuvaiskii <dimakuv@amazon.de>