matching: do not apply `runstatedir` backport when not needed
Previously, if we detected Autoconf v2.69, we always applied the
`runstatedir` backport patch. The reason for this was that many popular
distros (Fedora, Debian) ship with Autoconf v2.69 that has this patch
applied. However, this is not always the case, e.g. Arch Linux does not
do that. Thus, this commit adds a check: does the upstream archive
contain a hint that the `runstatedir` patch was applied? If yes, then
the patch is applied, otherwise vanilla Autoconf v2.69 is used.

Note: this whole dance with patching Autoconf is ugly. Basically, we
mimic the version of Autotools utilities as they are shipped in
different OS distros. Adding more and more "mimicking" code is
unsustainable, thus we must re-evaluate this approach.

Signed-off-by: Dmitrii Kuvaiskii <dimakuv@amazon.de>