[DONTMERGE] matching: use libtool v2.6.4 as shipped in Arch Linux
This is a hacky commit to explore what it takes to reproduce the
Autotools-generated files in the xz-5.2.5.tar.xz tarball (xz utils
software project). The xz tarball is known to be generated on (some
version) of Arch Linux and uses ArchLinux-specific version of libtool.
See how libtool is patched and built on Arch Linux here:
https://gitlab.archlinux.org/archlinux/packaging/packages/libtool/-/tree/2.4.6+42+gb88cebd5-16

With these changes, xz's autotools-generated files are reproducible. The
only exceptions are `Changelog` (in the official repo, this file exists
and has a dummy text, so our tool doesn't try to regenerate it) and `po`
(PO files, used for translations) which are non-reproducible generally.

Signed-off-by: Dmitrii Kuvaiskii <dimakuv@amazon.de>