rpm: improve error handling when downloading an SRPM
This is a cosmetic change. Previously, the code that handles errors
during download of a source package (SRPM) was inconsistent: sometimes
the code returned empty values instead of raising an exception, and some
errors were not logged at a proper (error) level. The resulting logs
were hard to parse; this commit improves the readability of logs a bit.

Signed-off-by: Dmitrii Kuvaiskii <dimakuv@amazon.de>