suggesting: recognize g-prefixed commit hashes from git-describe
Previously, `is_commit_hash()` only handled plain commit hashes but not
those with a "g" prefix that are commonly output by the git-describe
command. For example, git-describe might output "v1.0.0-5-gabcdef123456"
where "gabcdef123456" represents a commit hash with a "g" prefix. The
tool would fail to recognize such strings as commit hashes and
incorrectly classify them as regular version strings. This would lead in
turn to the tool not being able to find a commit in the git repo, and
ultimately failing to file-match the archive against a git repo state at
a specific commit.

This was found on Fedora 43 where the glibc package has an archive
`glibc-2.42-21-g7a8f3c6ee4.tar.xz`.

Signed-off-by: Dmitrii Kuvaiskii <dimakuv@amazon.de>